(Poor) attempt at trying to de-bloat the flag logic
