+++ src/main/java/de/t14d3/zones/permissions/flags/Flags.java
@@ -3,12 +3,12 @@ package de.t14d3.zones.permissions.flags;
 import de.t14d3.zones.Region;
 import de.t14d3.zones.permissions.Result;
 import de.t14d3.zones.utils.Types;
-import org.jetbrains.annotations.NotNull;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
+@SuppressWarnings("ClassWithTooManyFields")
 public class Flags {
 
     public static Flag FALLBACK;
@@ -40,59 +40,54 @@ public class Flags {
         FALLBACK = registerFlag(new Flag("fallback", "Fallback description"));
 
         BREAK = registerFlag(
-                new Flag("break", "Allows breaking blocks", new DefaultPlayerFlagHandler(false, Types.blocks())));
+                new Flag("break", "Allows breaking blocks", new DefaultFlagHandler(false, Types.blocks())));
         PLACE = registerFlag(
-                new Flag("place", "Allows placing blocks", new DefaultPlayerFlagHandler(false, Types.blocks())));
+                new Flag("place", "Allows placing blocks", new DefaultFlagHandler(false, Types.blocks())));
         INTERACT = registerFlag(
-                new Flag("interact", "Allows interacting", new DefaultPlayerFlagHandler(false, Types.all())));
+                new Flag("interact", "Allows interacting", new DefaultFlagHandler(false, Types.all())));
         CONTAINER = registerFlag(
                 new Flag("container", "Allows opening containers",
-                        new DefaultPlayerFlagHandler(false, Types.containers())));
+                        new DefaultFlagHandler(false, Types.containers())));
         REDSTONE = registerFlag(new Flag("redstone", "Allows interacting with redstone",
-                new DefaultPlayerFlagHandler(false, Types.redstone())));
+                new DefaultFlagHandler(false, Types.redstone())));
         ENTITY = registerFlag(new Flag("entity", "Allows interacting with entities",
-                new DefaultPlayerFlagHandler(false, Types.entities())));
+                new DefaultFlagHandler(false, Types.entities())));
         IGNITE = registerFlag(
                 new Flag("ignite", "Allows igniting tnt", new IgnitionFlagHandler(false, Types.blocks())));
         DAMAGE = registerFlag(
-                new Flag("damage", "Allows damaging entities", new DefaultPlayerFlagHandler(false, Types.entities())));
+                new Flag("damage", "Allows damaging entities", new DefaultFlagHandler(false, Types.entities())));
 
         GROUP = registerFlag(new Flag("group", "Add a group to the player"));
 
         SPAWN = registerFlag(new Flag("spawn", "Controls the spawning of entities",
-                new DefaultUniversalFlagHandler(true, Types.entities())));
+                new DefaultFlagHandler(true, Types.entities())));
         EXPLOSION = registerFlag(new Flag("explosion", "Controls the explosion of entities",
-                new DefaultUniversalFlagHandler(true, Types.entities())));
+                new DefaultFlagHandler(true, Types.entities())));
         CREATE = registerFlag(
                 new Flag("create", "Controls the creation of blocks through world events",
-                        new DefaultUniversalFlagHandler(true, Types.blocks())));
+                        new DefaultFlagHandler(true, Types.blocks())));
         DESTROY = registerFlag(
                 new Flag("destroy", "Controls the removal of blocks through world events",
-                        new DefaultUniversalFlagHandler(true, Types.blocks())));
+                        new DefaultFlagHandler(true, Types.blocks())));
         TRANSFORM = registerFlag(
                 new Flag("transform", "Controls the transformation of blocks into other blocks",
-                        new DefaultUniversalFlagHandler(true, Types.blocks())));
+                        new DefaultFlagHandler(true, Types.blocks())));
         SPREAD = registerFlag(
                 new Flag("spread", "Controls the spread of blocks through world events",
-                        new DefaultUniversalFlagHandler(true, Types.blocks())));
+                        new DefaultFlagHandler(true, Types.blocks())));
         RELOCATE = registerFlag(new Flag("relocate", "Controls the ability for blocks to change their location",
-                new DefaultUniversalFlagHandler(true, Types.blocks())));
+                new DefaultFlagHandler(true, Types.blocks())));
         HURT = registerFlag(new Flag("hurt", "Controls the ability for entities to be hurt",
-                new DefaultUniversalFlagHandler(true, Types.entities())));
+                new DefaultFlagHandler(true, Types.entities())));
 
         PHYSICS = registerFlag(new Flag("physics", "Controls the physics of blocks",
-                new DefaultUniversalFlagHandler(true, Types.blocks())));
+                new DefaultFlagHandler(true, Types.blocks())));
     }
 
-    private static final class IgnitionFlagHandler extends PlayerFlag {
+    private static final class IgnitionFlagHandler extends UnifiedFlag {
 
         public IgnitionFlagHandler(boolean defaultValue, List<String> validValues) {
             super(validValues, defaultValue);
-        }
-
-        @Override
-        public Result evaluate(Region region, String permission, String type, Object... optionals) {
-            return super.evaluate(region, permission, type, optionals);
         }
 
         @Override
@@ -118,7 +113,7 @@ public class Flags {
      * @return {@code true} if it was registered, {@code false} if it already exists
      * @see #registerFlag(String, String, boolean)
      */
-    public static boolean registerFlag(@NotNull String name, @NotNull String desc) {
+    public static boolean registerFlag(String name, String desc) {
         return registerFlag(name, desc, false);
     }
 
@@ -131,7 +126,7 @@ public class Flags {
      * @return {@code true} if it was registered, {@code false} if it already exists and {@code overwrite} is {@code false}
      * @see #registerFlag(String, String)
      */
-    public static boolean registerFlag(@NotNull String name, @NotNull String desc, boolean overwrite) {
+    public static boolean registerFlag(String name, String desc, boolean overwrite) {
         if (flags.stream().anyMatch(f -> f.name().equals(name)) && !overwrite) {
             return false;
         } else {
@@ -140,11 +135,11 @@ public class Flags {
         }
     }
 
-    public static Flag registerFlag(@NotNull Flag flag) {
+    public static Flag registerFlag(Flag flag) {
         return registerFlag(flag, false);
     }
 
-    public static Flag registerFlag(@NotNull Flag flag, boolean overwrite) {
+    public static Flag registerFlag(Flag flag, boolean overwrite) {
         if (flags.stream().anyMatch(f -> f.name().equals(flag.name()))) {
             if (overwrite) {
                 flags.removeIf(f -> f.name().equals(flag.name()));
@@ -170,6 +165,4 @@ public class Flags {
     public static Flag getFlag(String name) {
         return flags.stream().filter(f -> f.name().equals(name)).findFirst().orElse(FALLBACK);
     }
-
-
 }
